ci(docker): provide a Debian-based image which configures and builds somoclu for python
Previously, the Somoclu library, for python, was built based on Alpine Linux. Now we also provide a
Debian-based image with the idea that most python package distributers host on pypi both "source"
and "wheel" distributions (of their python packages). Wheel files being precompiled versions of the
code, usually target Debian operating systems, but not Alpine ones [python does not support it
(yet)]. That being said, invocations of "pip install" (the typical install command for a python
package) will look for wheels in pypi, speeding up installation, when running on Debian. So that is
why we provide the Debian-based image.

wip

ci(docker): provide a Debian-based image which configures and builds somoclu for python